changed photo-z's in spt_dndz_tab_f1_kappa_wd03.txt
Since the redshift for SPT0452-50 was updated, a new value for dust
temperature value was obtained using a modified blackbody fit, as
described in Reuter+20.  Because the photometric redshifts were
derived using a prior based on the dust temperature and SPT0452-50
was a cold outlier (z~2 Td~21K compared to z~5 and Td~70K), I
updated the Tdust distribution and re-fit photometric redshifts
for the SPT DSFG sample.  The majority of photo-z values agree with
 the originally published values within errors, except at low-z
where we lose some constraining power.