Updated SLED/faint lines values with SPT0452-50 at z~5
The 3mm spectrum for SPT0452-50 contained a single CO emission
line near the edge of the 3mm bandpass, which was originally
thought to be CO(3-2).  Since updating the spectroscopic redshift
to z=5.0160, this line is actually CO(6-5).  Similarly, we included
ATCA followup data showing an emission line thought to be CO(1-0)
in the original analysis, but is actually CO(2-1).

For simplicity, we omit the flux densities from both the observed
CO lines for SPT0452-50 from the faint line and CO SLED tables.
Because SPT0452-50 was originally thought to be a very cold (~21K)
source with high dust mass and we weighted the flux densities by
dust mass, SPT0452-50 had a strong influence on the CO values.
Despite the strong weighting, SPT0452-50 contributed little to the
stacked faint lines (faint = anything other than 12CO in this case)
and the new values agree with the published values within their
stated error bars.

Also, the ``new" ATCA sources presented in Reuter+23 were omitted
by mistake from the public github version of this table (but were
correctly included in the published version).  This data has now
been included in the online version of this table and is consistent
with the published version (except for the recent omission of
SPT0452-50, as explained above).